requirements.txt
numpy
torch
torchvision
scikit-learn
kornia
matplotlib
Pillow
tqdm
psutil
fastervit
